put useDataFetch in its own file
